refactor: split projection.rs (1,956 lines) into projection/ directory
Extract self-contained concerns from the monolithic projection module:

- projection/coverage.rs (462 lines): CoverageBoundary, ClampedPosition,
  PanoramaExtent - the precomputed pitch-indexed LUT for O(1) safe_clamp
- projection/virtual_camera.rs (101 lines): VirtualCamera orthonormal
  basis for yaw/pitch decomposition
- projection/geometry.rs (35 lines): point_in_polygon ray-casting utility
- projection/mod.rs: Projection trait, coordinate transforms, tests

All re-exports preserved - external code unchanged.
34 unit tests pass.
